Determine LINKER_SONAME.
Generating a shared libary usually involves:

cc -shared -Wl,-soname,lib.so.0

Or equivalent.  Except on Mac OS X, soname is install_name.

If the user doesn't override an input value of LINKER_SONAME, try
to determine which to use, and propogate that value to the
generated Makefile.configure.

Bump version as well.