Full crypt_newpass compat.
This adds two depending parts: _PASSWORD_LEN for Linux and
timingsafe_bcmp for others.  The compat embeds bcrypt() and the
bcrypt_newhash() functions, but does not expose them.

It also fixes the strndup regression not to raise a warning on Linux and
an incorrect definition of _GNU_SOURCE instead of _XOPEN_SOURCE to
expose crypt() from unistd instead of crypt.h on Linux.