Allow CC/AR/CFLAGS command-line override
To date, the CC/AR/CFLAGS variables could only be passed in by way of
the environment.  The environment would then be re-interpreted by `make`
and spit back out.  In the case of CFLAGS, the result was augmented by a
fixed set of arguments.

Allow CC/AR/CFLAGS to now be passed in as command-line arguments just
like the other variables.  The passed-in arguments override any
environment variables, and the result is filtered through `make` in the
same way as before.

The only difference with current behaviour is that if CFLAGS are either
passed in or given on the command-line, they are not augmented by the
fixed set of arguments used in the default case.  This gives operators
more control over exactly which arguments are used.