Add initial support for dylib (Mac OS X)
On Mac OS X, shared libraries are generated using the -dynamiclib flag
to the linker and the dylib suffix.  This is not to be confused with
dynamically-loaded libraries (i.e., via dlopen), which use the same
-shared semantics as on other Unix-like systems.

Provide some variables for testing whether the current systems supports
one or the other, and allow for overriding both.

This will set LINKER_SOFLAG to be the flag (-shared or -dynamiclib) and
LINKER_SOSUFFIX to be the suffix (so or dylib).