Parallelise configuration
This is the initial effort to parallelise the configuration sequence.
Parallel configuration is enabled by the -j flag.  By providing a
non-zero value to -j, that's the level of parallelism used.  By
providing a zero value, the number of processors is used.

The framework for parallelism is to have subprocesses write the
configuration log and result into a file.  The result is atomically
renamed to a "final" result form after all processing completes, then
exits.

The parent executes "wait" after starting the maximum number of
processes, then reaps the results in the "final" result form.  For each
reaped result, the number of processes is decremented.

All of the files are placed in a per-configuration temporary directory
that's blown away on exit.  To wit, don't try to remove every file
in-line and just depend on a "trap" at the end to remove everything.

While here, clean up the shell scripting: don't use superfluous ${curly}
braces, make sure to quote more variables, use a configlog() and log()
function instead of throwing around endless file descriptors, use $()
instead of backticks where appropriate, and remove comment horizontal
lines.